[#84] Indices size are uint32_t now, maximum value of vertices restriction

+++ UnicornEngine/video/source/Mesh.cpp
@@ -18,7 +18,7 @@ Mesh::~Mesh()
 {
 }
 
-void Mesh::SetMeshData(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices)
+void Mesh::SetMeshData(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices)
 {
     m_vertices = vertices;
     m_indices = indices;
@@ -38,7 +38,7 @@ const std::vector<Vertex>& Mesh::GetVertices() const
     return m_vertices;
 }
 
-const std::vector<uint16_t>& Mesh::GetIndices() const
+const std::vector<uint32_t>& Mesh::GetIndices() const
 {
     return m_indices;
 }
